Fix builds against glibc >= 2.34
Starting with glibc 2.34 SIGSTKSZ and a few others are no
longer constants, instead their values are computed at
runtime and are of type long which is incompatible with
previous size_t which was unsigned long.

Fix this by moving all invocations to run-time instead
of compile time.

BUG=b:235960683
TEST=Local builds against glibc 2.34.

Reviewed-on: https://chromium-review.googlesource.com/c/native_client/src/native_client/+/3714944
Auto-Submit: Adrian Ratiu <adrian.ratiu@collabora.corp-partner.google.com>
Commit-Queue: Mark Seaborn <mseaborn@chromium.org>

+++ src/trusted/service_runtime/nacl_signal_frame_test.c
@@ -33,7 +33,7 @@
  */
 
 int main(void) {
-  size_t stack_size = SIGSTKSZ;
+  long stack_size = SIGSTKSZ;
   char *stack;
   stack_t st;
   int rc;
